Make `@DeprecationSummary` have an effect in the Discussion section
When specifying the @DeprecationSummary directive in a symbol extension markdown file, the directive wasn't taking effect if the symbol had multiple lines of documentation comments.

This was happening because `DocumentationMarkup.deprecation` was only being populated in the abstract. I've now updated it to take an effect in the Discussion section as well.

Resolves rdar://70056350.