Update `BlockDirective` extensions (#983)
* Move `directivesRemovedFromContent` to `BlockDirective`

While looking at https://github.com/swiftlang/swift-docc/pull/982, I found that there's a better place to keep the list of directive names which we consider removed from content -- we already had a file with lists of directive names so it'll be clearer to keep all of them in the same file.

* Add missing directive names to `BlockDirective.allKnownDirectiveNames`

A lot of directive names were missing from `BlockDirective.allDirectiveNames`. I've updated the array to include all the missing directives which I could find.

This property is used to throw warnings about directives which are not supported in symbol source documentation.

* Remove duplicated `directivesRemovedFromContent` definition

`directivesRemovedFromContent` was being defined in both `DocumentationMarkup` and `BlockDirective`, removed the duplicate definition.