Support for Possible Values section.
Support a new markdown tag, `possibleValues`, and a new `Possible Values` render section to document possible values extracted from the SymbolGraph.

If a documented possible value does not correspond with the values from the SymbolGraph, it is dropped. If no possible values are documented in the markdown, the possible values are rendered under the `Attributes`render section, and the `Possible Values` render section is not displayed.

rdar://123262314