Allow default availability entries wothout version number.
This change allows defining entires in the Default acailability section of the `Info.plist` file but without specifiying an introduced version.

This change aims to not enforce a potentiall wrong version in symbols with unknown introduced version number, but that we still want to convey they are available in a certain platform. Also, allows to convey that a framework is available in other platforms, such as Linux.

rdar://132980711