Experimental flag added for availability items without version.
In order to hide the availability items without version feature (`f361d5f`) a new experimental flag is added `ExperimentalAvailabilityItemsWithoutVersion`.

This flag is set to false by default, and will be used while we discover and test the possible implications of allowing availability items without version as part of the framework default availability.