Always set role even if automatic role headings are disabled
Fixes a bug where disabling automatic role headings with directive `@AutomaticTitleHeading` would also result in the node's role being lost.

We always want to set the node's role as part of the node's metadata, regardless of any title settings.
Some unit tests were added to ensure that the interaction between `AutomaticTitleHeading`, `.metadata.roleHeading` and `.metadata.role` is all working as expected.

Additional notes:
-----------------
We still want the role heading to be set if it has been overridden with `@TitleHeading`, regardless of the value of `@AutomaticTitleHeading`.
Only automatic title headings should be disabled, not manually curated ones.
The same applies for `@PageKind`, which manually overrides the role and role heading.

Fixes rdar://128609308.