Added a precondition for declaring a node as beta.
For a node to be declared as beta it has to contain at least a single availability item that contains all the already defined characteristics.

This code adds a precondition where at least the node needs to have one availability item to be considered as potentially in beta.

Before this change, if at somewhere in the code, the availability items got removed from the node, the node would be considered as beta.

rdar://137762221