Make `NavigatorIndex.Builder` ignore language variants when requested
Clients might want the Navigator Index to not preemptively add an entry
for nodes that have a language variant trait. For example, a client
might not have a renderer cable of applying language variants, in this
instance it doesn't make sense to generate a navigator hierarchy for
other languages.

rdar://138183564