Allow availability items with deprecated version and no introduced versions.
There can be symbols that define an explicit default avaialblility verion in the symbol graph but don't define an introduced version. Before this change DocC would drop the availability item entirely.

This change allows this case without droping the item.

This is needed since the `@available` attribute allows defining only the deprecated version.

    rdar://138034711