Filter out oboleted availability items.
Obsoleted availbaility is not something that we want to display in the final documentation of a symbol.
Before the change made on the first commit this logic was not needed since the availability items that surfaced had the requirement of an introduced version, but since the introduction of versionless availability items this is not the case anymore, but we don't want this information, that is represented in the SGF in the following way:

```
"availability":[
	{
		"domain":"platform name",
		"obsoleted":{"major":13,"minor":0}
	}
]
```
to be rendered in the final documentation.