Allow availability items with only deprecated versions. (#1079)
* Allow availability items with deprecated version and no introduced versions.

There can be symbols that define an explicit default avaialblility verion in the symbol graph but don't define an introduced version. Before this change DocC would drop the availability item entirely.

This change allows this case without droping the item.

This is needed since the `@available` attribute allows defining only the deprecated version.

    rdar://138034711

* Filter out oboleted availability items.

Obsoleted availbaility is not something that we want to display in the final documentation of a symbol.
Before the change made on the first commit this logic was not needed since the availability items that surfaced had the requirement of an introduced version, but since the introduction of versionless availability items this is not the case anymore, but we don't want this information, that is represented in the SGF in the following way:

```
"availability":[
	{
		"domain":"platform name",
		"obsoleted":{"major":13,"minor":0}
	}
]
```
to be rendered in the final documentation.