Add `@AlternateRepresentation` directive
Adds a new child directive to `@Metadata` which can be used in a symbol extension file by specifying the link to another symbol:
```swift
@Metadata {
  @AlternateRepresentation(``MyClass/property``)
}
```

External links are also supported, as long as they're quoted:
```swift
@Metadata {
  @AlternateRepresentation("doc://com.example/documentation/MyClass/property")
}
```

The intent of this directive is to define an alternate language representation for the current symbol, such that both symbols are considered to be alternate representations of the same symbol.

Ideally two symbols which are equivalent would have the same USR and be resolved as the same symbol by the compiler, but this is not always the case. For the cases in which it is not feasible to change the source code to have them as one symbol, the `@AlternateRepresentation` directive can be used to manually link them as variants of each other.

Discussion:
----------

A mutable topic reference type was chosen as the type for parsing&storing the link
so that it can be parsed as an unresolved link at the directive parsing stage,
and then later be updated to a resolved link at a later stage when the documentation context is resolving all links.

A parsing failure diagnostic is returned if the link is invalid in any way:
```
AlternateRepresentation expects an argument for an unnamed parameter that's convertible to 'TopicReference'
 --> SynonymSample.docc/Symbol.md:4:31-4:37
2 |
3 | @Metadata {
4 +     @AlternateRepresentation("doc://")
5 | }
6 |

```

This commit adds the directive itself, but doesn't hook it up to other parts of SwiftDocC. Subsequent commits will add:
- link resolution
- rendering logic

Alternatives considered:
-----------------------
Considered other names such as `@Synonym`, `@Counterpart`, `@AlternativeDeclaration` and `@VariantOf`.
In the end disqualified these as being confusing, and chose `@AlternateRepresentation` for being the one which strikes the best balance between readable and closeness to the technical term for this concept.