Resolve topic references in `@AlternateRepresentation`
Adds logic in `DocumentationContext` which will resolve the references inside the alternate representation directive.

The same logic is used as for resolving all other links.
This is done outside the usual ReferenceResolver visit of the semantic object, because `Symbol` objects don't have access to the node metadata, where the unresolved link resides.

If the link cannot be resolved, the usual diagnostic is emitted:
```
warning: 'MissingSymbol' doesn't exist at '/Synonyms'
 --> SynonymSample.docc/SymbolExtension2.md:4:19-4:32
2 |
3 | @Metadata {
4 +     @AlternateRepresentation(``Synonyms/MissingSymbol``)
5 | }
```