Support `@DeprecationSummary` in doc comments
Adds support for authoring `@DeprecationSummary` messages in
documentation comments. This is useful when you want to provide a
formatted documentation summary for an API that's deprecated, but don't
want to create a documentation extension file in order to do just that.

It turns out that DocC already supported this, but emitted a warning
whenever you did so. This commit removes that warning.