Fix availability logic refactor (#1065)
Refactor symbol availability logic.

This refactor was initiated by the need of stop using the module operating system name as an availability item.

This change required a big refactoring of the logic, and now the symbol graph availability loading logic is like the following:
1. Every symbol graph is loaded and the symbols get only the availability information that's marked in the SDK.
2. Once all the SGFs are oaded, whe iterate over the symbols and we add the fallbkac and default availability if applies.

* Refactor unit tests for testing symbol availability.

* Refactor fallback platforms data structure.

Instead of making a `[(FallbackPlatform, InheritedPlaform)].
The new struture to hold this information is in the way of `[InheritedPlatfor:[FallbackPlatforms]]`.