adds a base symbolgraph for Snippets test fixture
Supports fixing an issue confusing snippets as both "main" and "extension"
symbolgraphs for a particular module. The fix in swift-docc-symbolkit
fails this specific fixture since it doesn't currently have a "main"
graph, even empty as this one, to extend from. With this in place, the
tests fully pass for properly resolving snippets symbolgraphs reliably.