Allow relative references to images across bundle identifiers
Fixes an issue where roundtripping a `URLReference` would yield different results depending on the bundle identifier of the catalog.

Paths to images and other downloadable content would always get rendered into JSON with the bundle identifier of the current bundle, discarding the original path, regardless of whether they already contained a reference to a different bundle.

This causes issues when the resolved information from an out of process resolver contains a relative path, e.g. `/images/com.example.bundle/image.png`, as the path would always be rewritten to be the catalog's bundle identifier.

This fix makes it such that if the path is already relative to the base URL of the reference type, the URL is left unchanged.

Fixes rdar://146108301.