Add mapping from documentation kind to symbol kind
There is a mapping from `SymbolGraph.Symbol.KindIdentifier` to `DocumentationNode.Kind`, but not the other way around; this commit adds one.

For documentation kinds which both map to the same symbol kind, this has been expressed as:
- both documentation kinds are converted to the same symbol kind in `symbolKind(for:)`.
- a specific documentation kind is chosen as the default mapping in `kind(forKind:)`

For example, for `DocumentationNode.Kind. typeConstant` [1]:
- both `symbolKind(for: .typeConstant)` and `symbolKind(for: .typeProperty)` return `.typeProperty`
- `kind(forKind: .typeProperty)` returns `.typeProperty`

This function will be used to map and external entity's documentation kind to a symbol kind, so that that information can be populated as part of the navigator.

## Verification:

This has been verified by testing the round trip conversion of all symbol kinds, and all documentation kinds which are symbols.

[1]: https://github.com/swiftlang/swift-docc/blob/65aaf926ec079ddbd40f29540d4180a70af99e5e/Sources/SwiftDocC/Model/Kind.swift#L151