RenderNodeTranslator.visitSymbol filters out availability items which have a missing or invalid domain (platform name).
Swift developers can add global availability messages with no specific
platform information, for example like this:

@available(*, deprecated, message:
  """
  Don't use this function; call some other function instead.
  """
)

In this scenario, the Swift compiler adds an availability item with the
message but no domain (platform name) to the symbol graph file, like
this:

{
  "domain": null,
  "isUnconditionallyDeprecated": true,
  "message": "Don't use this function; call some other function instead."
}

To avoid displaying an empty/null platform in the DocC Render user
interface, the RenderNodeTranslator needs to filter out these records.
Additionally, AvailabilityRenderOrder.compare always returns false for
availability items that have no name, leading to an unpredictable sort
order when any of the items is missing a platform name, even for
availability items that had a valid platform name.